# gate/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/redis/hash": {
            "get": {
                "description": "Hash",
                "tags": [
                    "Redis"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/redis/list": {
            "get": {
                "description": "List",
                "tags": [
                    "Redis"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/redis/set": {
            "get": {
                "description": "Set",
                "tags": [
                    "Redis"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/redis/sortedSet": {
            "get": {
                "description": "SortedSet",
                "tags": [
                    "Redis"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/redis/string": {
            "get": {
                "description": "String",
                "tags": [
                    "Redis"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "data",
                        "name": "data",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/byFormData": {
            "post": {
                "description": "取得用戶",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "姓名",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/byJson": {
            "post": {
                "description": "ByJson",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "description": "json string",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/byParams": {
            "get": {
                "description": "取得用戶",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用戶id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/byRaw": {
            "post": {
                "description": "Raw data",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "description": "json string",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/{id}": {
            "get": {
                "description": "取得用戶",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用戶id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/type/{id}/{name}": {
            "get": {
                "description": "ByJson",
                "tags": [
                    "API類型"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用戶id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "用戶姓名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/user/:id": {
            "get": {
                "description": "取得用戶",
                "tags": [
                    "取得用戶"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用戶id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\",\"method\",\"path\",\"id\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}